refactor: add types file.
